Sources/Server/requirements.txt
flask
RPi.GPIO
requests
